Reorgnized utils unit testing.
